attendance_backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
